More Comments, and moving methods into their own classes
